feat(finder): Overhaul the code and make it a library
This is a big commit, but it builds hugely on @benv123's big PR #23. What I've
done here is:

1. Try to make it into a good Python package. I think this should be just about
   good enough for PyPi, once the readme and a few other random things are
   done.

2. Take the code Ben wrote and make it a bit more mature. I did this by adding
   tests, and doing a huge refactor to use a filtering pattern instead of
   deeply-nested for loops.

The basic idea now though is that you can do something like:

```
import xray
bad_redactions = xray.inspect("some_file/path.pdf")
```

When you do that, the result you get is a dictionary with the key being the
page number and the value being a list of bad redactions including their
location and text.

[pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

feat(finder): Overhaul the code and make it a library

This is a big commit, but it builds hugely on @benv123's big PR #23. What I've
done here is:

1. Try to make it into a good Python package. I think this should be just about
   good enough for PyPi, once the readme and a few other random things are
   done.

2. Take the code Ben wrote and make it a bit more mature. I did this by adding
   tests, and doing a huge refactor to use a filtering pattern instead of
   deeply-nested for loops.

The basic idea now though is that you can do something like:

```
import xray
bad_redactions = xray.inspect("some_file/path.pdf")
```

When you do that, the result you get is a dictionary with the key being the
page number and the value being a list of bad redactions including their
location and text.