feat(pdfs): Add support for complex PDFs
This handles a number of difficult corner cases we've run into:

 - Clipping paths
 - Complex drawings with complex winding rules

The general idea now is to over-include these (which we were doing
before), and to then filter them out by looking at the pixels of
the rendered redaction.

If there's a lot of variety in the rendered redaction pixels, we
can assume it's not a bad redaction, and it's a false positive.

This also adds support for redactions hidden under rectangles of
any color.